+++ mainuttt.js
@@ -1,5 +1,5 @@
 var script = require('./uttt');
 
-script.mlSimulateGames(parseInt(process.argv[3]), parseInt(process.argv[4]), process.argv[2]);
+script.mlSimulateGames(parseInt(process.argv[3]), parseFloat(process.argv[4]), process.argv[2]);
